LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/mq-invalid-media-type-005-expected.txt


FAIL syntactical MQ keywords used as media types are a syntax error assert_equals: expected "not all" but got "and"
